Fix motif style knowledge detection for U33
Since the introduction of account wide achievements in update 33 all characters share the
achievement progress for motif style achievements. This leads to several issues regarding
the detection, the current character knows a certain motif chapter or not.
The new way to check for knowledge of the chapter is to check, if the associated style
book is known.

The API update introduced a new function
`GetAchievementLinkedBookCollectionId(achievementId)` which is used to bridge the gap
to the book collection. This allows us to continue using the achievement IDs instead of
listing up all book collections manually.