Fix highlighting of tracked study lines in U34
With the introduction of the new TLC rules in update 34, some
functions seem to have been removed from texture controls
(or were made inaccessible).

This lead to the following UI error:
user:/AddOns/CraftStoreFixedAndImproved/CraftStore.lua:824:
function expected instead of nil

This fix simply removes the additional highlighting from the
study line header that consisted of a red border and
background. The red hue change of the study line's
header icons still remains however, which should be enough
of an indicator to show if the line is fully tracked.