Add filter on startup to remove unpublished styles and furnishings
In order to avoid the additional effort to check which styles made it
from a PTS to the live version, an initial filtering has been added
that checks that each style has an existing reference item and
a linked achievement (except crown store styles).

Similarly, the list of furnishing recipies is filtered and all unknown
recipies are removed from the list as well.

This change should also allow publishing now allow for the
publishing of CrownStore updates before the release
of a new major ESO update, provided it does not contain any
breaking changes regarding the API.