Bring windows to the top when opening them
With update 34 and the changes to the GUI rendering, windows that
were opened (switched from hidden to visible) were placed behind
the CraftStore main window.

This change brings all TopLevelControls to the top when they
are shown through their associated buttons. Afterwards, the layering
is controled by window focus, i.e. when you click a window, it is
brought to the top as expected.