Fix misaligned tracking markers
A previous bugfix that addressed a change to the way
child controls are enumerated update 34 included
a change to the start index for the trait controls within a
research line. That fix introduced an off-by-one error
when selecting the tracking information for the trait
which lead to CraftStore marking the wrong trait cell
after clicking it.

This change now fixes this off-by-one error by applying
the required offset to align with the tracking table.