Fix broken recipe sorting for russian clients by searching for alternative spellings of resource names
Due to some quirks in the russian language or cyrillic alphabet, the resource names do not appear
the same way in recipe descriptions as they do in other parts of the game, leading to missmatching
when sorting the recipes in the different categories.

This change introduces a list of alternative resource names in the language file which will be used
for additional matching attempts if it is present.

Changelog: fixed