Disable hiding of CraftstoreCook when clicking the fillet button in gamepad mode
As the gamepad UI's control names differ from the keyboard UI, the hooks for the tab
selector buttons failed to register, leading to UI errors for gamepad users.

With this change the hooks will only be registered for keyboard users and CSCook
will remain opened along the gamepad cooking menu, as it already did before with
all other functionality.

Changelog: fixed